Move DPS tree code to DPS folder
