+++ src/Nodes/DPS/DpsResourceTreeItem.ts
@@ -3,7 +3,7 @@
 
 import { IotDpsModels } from "azure-arm-deviceprovisioningservices";
 import { AzureParentTreeItem, AzureTreeItem, TreeItemIconPath } from "vscode-azureextensionui";
-import { TreeUtils } from "../Utility/treeUtils";
+import { TreeUtils } from "../../Utility/treeUtils";
 
 // Represents a DPS resource
 export class DpsResourceTreeItem extends AzureTreeItem {
