+++ src/extension.ts
@@ -13,8 +13,8 @@ import { Constants, DistributedSettingUpdateType } from "./constants";
 import { DeviceTree } from "./deviceTree";
 import { Executor } from "./executor";
 import { DeviceNode } from "./Nodes/DeviceNode";
-import { DpsAccountTreeItem } from "./Nodes/DpsAccountTreeItem";
-import { DpsResourceTreeItem } from "./Nodes/DpsResourceTreeItem";
+import { DpsAccountTreeItem } from "./Nodes/DPS/DpsAccountTreeItem";
+import { DpsResourceTreeItem } from "./Nodes/DPS/DpsResourceTreeItem";
 import { EventHubItemNode } from "./Nodes/Endpoints/EventHubItemNode";
 import { IoTHubAccountTreeItem } from "./Nodes/IoTHub/IoTHubAccountTreeItem";
 import { IoTHubResourceTreeItem } from "./Nodes/IoTHub/IoTHubResourceTreeItem";
